Aggregate FastQC Reports for Multiple Samples

Aggregate multiple FastQC reports into a data frame.

qc_aggregate(qc.dir = ".", progressbar = TRUE)

# S3 method for qc_aggregate
summary(object, ...)

qc_stats(object)

Arguments

qc.dir
path to the FastQC result directory to scan.

progressbar
logical value. If TRUE, shows a progress bar.

object
an object of class qc_aggregate.

...
other arguments.

Value

qc_aggregate() returns an object of class qc_aggregate which is a (tibble) data frame with the following column names:

sample: sample names

module: fastqc modules

status: fastqc module status for each sample

tot.seq: total sequences (i.e.: the number of reads)

seq.length: sequence length

pct.gc: % of GC content

pct.dup: % of duplicate reads

summary: Generates a summary of qc_aggregate. Returns a data frame with the following columns:

module: fastqc modules

nb_samples: the number of samples tested

nb_pass, nb_fail, nb_warn: the number of samples that passed, failed and warned, respectively.

failed, warned: the name of samples that failed and warned, respectively.

qc_stats: returns a data frame containing general statistics of fastqc reports. columns are: sample, pct.dup, pct.gc, tot.seq and seq.length.

Functions

qc_aggregate: Aggregate FastQC Reports for Multiple Samples

qc_stats: Creates general statistics of fastqc reports.

Examples

# Demo QC dir
qc.dir <- system.file("fastqc_results", package = "fastqcr")
qc.dir

#> [1] "/Users/kassambara/Documents/R/MyPackages/fastqcr/inst/fastqc_results"

# List of files in the directory
list.files(qc.dir)

#> [1] "S1_fastqc.zip" "S2_fastqc.zip" "S3_fastqc.zip" "S4_fastqc.zip"
#> [5] "S5_fastqc.zip"

# Aggregate the report
qc <- qc_aggregate(qc.dir, progressbar = FALSE)
qc

#> [90m# A tibble: 60 x 7[39m
#>    sample module                       status tot.seq seq.length pct.gc pct.dup
#>  [90m*[39m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m                        [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m   [3m[90m<chr>[39m[23m       [3m[90m<dbl>[39m[23m   [3m[90m<dbl>[39m[23m
#> [90m 1[39m S1     Basic Statistics             PASS   [31mNA[39m      [31mNA[39m             [31mNA[39m    17.2
#> [90m 2[39m S1     Per base sequence quality    PASS   [31mNA[39m      [31mNA[39m             [31mNA[39m    17.2
#> [90m 3[39m S1     Per tile sequence quality    PASS   [31mNA[39m      [31mNA[39m             [31mNA[39m    17.2
#> [90m 4[39m S1     Per sequence quality scores  PASS   [31mNA[39m      [31mNA[39m             [31mNA[39m    17.2
#> [90m 5[39m S1     Per base sequence content    FAIL   [31mNA[39m      [31mNA[39m             [31mNA[39m    17.2
#> [90m 6[39m S1     Per sequence GC content      WARN   [31mNA[39m      [31mNA[39m             [31mNA[39m    17.2
#> [90m 7[39m S1     Per base N content           PASS   [31mNA[39m      [31mNA[39m             [31mNA[39m    17.2
#> [90m 8[39m S1     Sequence Length Distribution WARN   [31mNA[39m      [31mNA[39m             [31mNA[39m    17.2
#> [90m 9[39m S1     Sequence Duplication Levels  PASS   [31mNA[39m      [31mNA[39m             [31mNA[39m    17.2
#> [90m10[39m S1     Overrepresented sequences    PASS   [31mNA[39m      [31mNA[39m             [31mNA[39m    17.2
#> [90m# ... with 50 more rows[39m

# Generates a summary of qc_aggregate
summary(qc)

#> [90m# A tibble: 12 x 7[39m
#> [90m# Groups:   module [?][39m
#>    module            nb_samples nb_fail nb_pass nb_warn failed      warned     
#>    [3m[90m<chr>[39m[23m                  [3m[90m<dbl>[39m[23m   [3m[90m<dbl>[39m[23m   [3m[90m<dbl>[39m[23m   [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m       [3m[90m<chr>[39m[23m      
#> [90m 1[39m Adapter Content            5       0       5       0 [31mNA[39m          [31mNA[39m         
#> [90m 2[39m Basic Statistics           5       0       5       0 [31mNA[39m          [31mNA[39m         
#> [90m 3[39m Kmer Content               5       0       5       0 [31mNA[39m          [31mNA[39m         
#> [90m 4[39m Overrepresented …          5       0       5       0 [31mNA[39m          [31mNA[39m         
#> [90m 5[39m Per base N conte…          5       0       5       0 [31mNA[39m          [31mNA[39m         
#> [90m 6[39m Per base sequenc…          5       5       0       0 S1, S2, S3… [31mNA[39m         
#> [90m 7[39m Per base sequenc…          5       0       5       0 [31mNA[39m          [31mNA[39m         
#> [90m 8[39m Per sequence GC …          5       2       0       3 S3, S4      S1, S2, S5 
#> [90m 9[39m Per sequence qua…          5       0       5       0 [31mNA[39m          [31mNA[39m         
#> [90m10[39m Per tile sequenc…          5       0       5       0 [31mNA[39m          [31mNA[39m         
#> [90m11[39m Sequence Duplica…          5       0       5       0 [31mNA[39m          [31mNA[39m         
#> [90m12[39m Sequence Length …          5       0       0       5 [31mNA[39m          S1, S2, S3…

# General statistics of fastqc reports.
qc_stats(qc)

#> [90m# A tibble: 5 x 5[39m
#>   sample pct.dup pct.gc tot.seq seq.length
#>   [3m[90m<chr>[39m[23m    [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m   [3m[90m<chr>[39m[23m     
#> [90m1[39m S1        17.2     [31mNA[39m [31mNA[39m      [31mNA[39m        
#> [90m2[39m S2        15.7     [31mNA[39m [31mNA[39m      [31mNA[39m        
#> [90m3[39m S3        22.1     [31mNA[39m [31mNA[39m      [31mNA[39m        
#> [90m4[39m S4        19.9     [31mNA[39m [31mNA[39m      [31mNA[39m        
#> [90m5[39m S5        18.2     [31mNA[39m [31mNA[39m      [31mNA[39m