Read a collection of FastQC data files

A wrapper function around qc_read to read multiple FastQC data files at once.

qc_read_collection(files, sample_names, modules = "all", verbose = TRUE)

Arguments

files

A character vector of paths to the files to be imported.

sample_names

A character vector of length equals that of the first argument files

modules

Character vector containing the names of FastQC modules for which you want to import/inspect the data. Default is all. Allowed values include one or the combination of:

"Summary",

"Basic Statistics",

"Per base sequence quality",

"Per tile sequence quality",

"Per sequence quality scores",

"Per base sequence content",

"Per sequence GC content",

"Per base N content",

"Sequence Length Distribution",

"Sequence Duplication Levels",

"Overrepresented sequences",

"Adapter Content",

"Kmer Content"

Partial match of module names allowed. For example, you can use modules = "GC content", instead of the full names modules = "Per sequence GC content".

verbose

logical value. If TRUE, print filename when reading.

Value

A list of tibbles containing the data of specified modules form each file.

Author

Mahmoud Ahmed, mahmoud.s.fahmy@students.kasralainy.edu.eg

Examples

# extract paths to the demo files
qc.dir <- system.file("fastqc_results", package = "fastqcr")
qc.files <- list.files(qc.dir, full.names = TRUE)

# read all modules in all files
qc <- qc_read_collection(qc.files, sample_names = paste('S', 1:5, sep = ''))
#> Reading: /private/var/folders/xm/8p6yj4bj6s57n4v_51714lwm0000gp/T/RtmpT6jSz8/temp_libpatha9b35de1589b/fastqcr/fastqc_results/S1_fastqc.zip
#> Warning: Missing column names filled in: 'X1' [1]
#> Warning: Missing column names filled in: 'X1' [1]
#> Reading: /private/var/folders/xm/8p6yj4bj6s57n4v_51714lwm0000gp/T/RtmpT6jSz8/temp_libpatha9b35de1589b/fastqcr/fastqc_results/S2_fastqc.zip
#> Warning: Missing column names filled in: 'X1' [1]
#> Warning: Missing column names filled in: 'X1' [1]
#> Reading: /private/var/folders/xm/8p6yj4bj6s57n4v_51714lwm0000gp/T/RtmpT6jSz8/temp_libpatha9b35de1589b/fastqcr/fastqc_results/S3_fastqc.zip
#> Warning: Missing column names filled in: 'X1' [1]
#> Warning: Missing column names filled in: 'X1' [1]
#> Reading: /private/var/folders/xm/8p6yj4bj6s57n4v_51714lwm0000gp/T/RtmpT6jSz8/temp_libpatha9b35de1589b/fastqcr/fastqc_results/S4_fastqc.zip
#> Warning: Missing column names filled in: 'X1' [1]
#> Warning: Missing column names filled in: 'X1' [1]
#> Reading: /private/var/folders/xm/8p6yj4bj6s57n4v_51714lwm0000gp/T/RtmpT6jSz8/temp_libpatha9b35de1589b/fastqcr/fastqc_results/S5_fastqc.zip
#> Warning: Missing column names filled in: 'X1' [1]
#> Warning: Missing column names filled in: 'X1' [1]

# read a specified module in all files
qc <- qc_read_collection(qc.files, 
    sample_names = paste('S', 1:5, sep = ''),
    modules = "Per base sequence quality")
#> Reading: /private/var/folders/xm/8p6yj4bj6s57n4v_51714lwm0000gp/T/RtmpT6jSz8/temp_libpatha9b35de1589b/fastqcr/fastqc_results/S1_fastqc.zip
#> Reading: /private/var/folders/xm/8p6yj4bj6s57n4v_51714lwm0000gp/T/RtmpT6jSz8/temp_libpatha9b35de1589b/fastqcr/fastqc_results/S2_fastqc.zip
#> Reading: /private/var/folders/xm/8p6yj4bj6s57n4v_51714lwm0000gp/T/RtmpT6jSz8/temp_libpatha9b35de1589b/fastqcr/fastqc_results/S3_fastqc.zip
#> Reading: /private/var/folders/xm/8p6yj4bj6s57n4v_51714lwm0000gp/T/RtmpT6jSz8/temp_libpatha9b35de1589b/fastqcr/fastqc_results/S4_fastqc.zip
#> Reading: /private/var/folders/xm/8p6yj4bj6s57n4v_51714lwm0000gp/T/RtmpT6jSz8/temp_libpatha9b35de1589b/fastqcr/fastqc_results/S5_fastqc.zip