Build a QC Report

Create an HTML file containing FastQC reports of one or multiple files. Inputs can be either a directory containing multiple FastQC reports or a single sample FastQC report.

qc_report(qc.path, result.file, experiment = NULL, interpret = FALSE,
  template = NULL, preview = TRUE)

Arguments

qc.path
path to the FastQC reports. Allowed values include:

A path to a directory containing multiple zipped FastQC reports,

Or a single sample zipped FastQC report. Partial match is allowed for sample name.

result.file
path to the result file prefix (e.g., path/to/qc-result). Don't add the file extension.

experiment
text specifying a short description of the experiment. For example experiment = "RNA sequencing of colon cancer cell lines".

interpret
logical value. If TRUE, adds the interpretation of each module.

template
a character vector specifying the path to an Rmd template. file.

preview
logical value. If TRUE, shows a preview of the report.

Examples

# NOT RUN {
# Demo QC Directory
qc.path <- system.file("fastqc_results", package = "fastqcr")
qc.path

# List of files in the directory
list.files(qc.path)

# Multi QC report
qc_report(qc.path, result.file = "~/Desktop/result")

# QC Report of one sample with plot interpretation
 qc.file <- system.file("fastqc_results", "S1_fastqc.zip", package = "fastqcr")
 qc_report(qc.file, result.file = "~/Desktop/result",
   interpret = TRUE)
# }